+++ modules/App/assets/js/app/ui.js
@@ -1,22 +1,6 @@
 
 let uuid = 0;
 
-function setHighestZindex(element) {
-
-    let highestZindex = parseInt(window.getComputedStyle(element).zIndex),
-        offsetParent = element.offsetParent || document.body,
-        zIndex;
-
-    Array.from(offsetParent.children).forEach((node) => {
-        zIndex = parseInt(window.getComputedStyle(node).zIndex) || 0;
-        if (zIndex > highestZindex) highestZindex = zIndex;
-    });
-
-    element.style.zIndex = highestZindex + 1;
-}
-
-window.setHighestZindex = setHighestZindex;
-
 export default {
 
     notify: function (message, status, timeout) {
@@ -97,8 +81,6 @@ export default {
 
             setTimeout(() => {
 
-                setHighestZindex(offcanvas);
-
                 let ele = offcanvas.querySelector('[autofocus]');
 
                 if (ele) {
@@ -141,8 +123,6 @@ export default {
             dialog.__show();
 
             setTimeout(() => {
-
-                setHighestZindex(dialog);
 
                 let ele = dialog.querySelector('[autofocus]');
 
@@ -276,8 +256,6 @@ export default {
             popout.__show();
 
             setTimeout(() => {
-
-                setHighestZindex(popout);
 
                 let ele = popout.querySelector('[autofocus]');