+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - make /api/system/healthcheck available without token
 - Update Vue to 3.4.26
 - Improve system health check
+- Fix possible dialog, offcanvas and popup z-index conflicts
 
 ## 2.8.3 (2024-04-08)